+++ .github/ISSUE_TEMPLATE/new-device-request.yml
@@ -13,7 +13,6 @@ body:
       label: Device
       options:
         - Choose a device...
-        - Delta Pro 3
         - Smart Home Panel
         - Power Kits
       default: 0
