+++ src/accessories/powerstream/interfaces/powerStreamHttpApiContracts.ts
@@ -22,8 +22,8 @@ export interface Heartbeat {
   // Settings
   supplyPriority?: SupplyPriorityType; //Power supply priority: 0: prioritize power supply; 1: prioritize power storage (read/write)
   permanentWatts?: number; //0.1 W; Custom load power (power of loads not connected to smart plugs),  0–600W (read/write)
-  upperLimit?: number; //Upper Charge limit, 1-30% (read/write)
-  lowerLimit?: number; //Discharge limit, 70-100% (read/write)
+  upperLimit?: number; //Upper Charge limit, 70-100% (read/write)
+  lowerLimit?: number; //Discharge limit, 1-30% (read/write)
 }
 
 export interface PowerStreamAllQuotaData {
